Add files via upload
This is an excel file with the number of vans and their destinations and the distances of each of those locations from the university. This will work as the data set for our project. Some of the points, their respective distances and the departure times are yet to be added.